Extract static rendering in to a module
