+++ pdoc/cli.py
@@ -1,14 +1,12 @@
 import argparse
 
-import codecs
-import os
-import os.path
 import sys
 
-import pdoc.web
 import pdoc.doc
 import pdoc.extract
 import pdoc.render
+import pdoc.static
+import pdoc.web
 
 version_suffix = "%d.%d" % (sys.version_info[0], sys.version_info[1])
 
@@ -108,60 +106,6 @@ def _eprint(*args, **kwargs):
     print(*args, **kwargs)
 
 
-def module_file(args, m):
-    mbase = os.path.join(args.html_dir, *m.name.split("."))
-    if m.is_package():
-        return os.path.join(mbase, pdoc.render.html_package_name)
-    else:
-        return "%s%s" % (mbase, pdoc.render.html_module_suffix)
-
-
-def quit_if_exists(args, m):
-    def check_file(f):
-        if os.access(f, os.R_OK):
-            _eprint("%s already exists. Delete it or run with --overwrite" % f)
-            sys.exit(1)
-
-    if args.overwrite:
-        return
-    f = module_file(args, m)
-    check_file(f)
-
-    # If this is a package, make sure the package directory doesn't exist
-    # either.
-    if m.is_package():
-        check_file(os.path.dirname(f))
-
-
-def html_out(args, m, html=True):
-    f = module_file(args, m)
-    if not html:
-        f = module_file(args, m).replace(".html", ".md")
-    dirpath = os.path.dirname(f)
-    if not os.access(dirpath, os.R_OK):
-        os.makedirs(dirpath)
-    try:
-        with codecs.open(f, "w+", "utf-8") as w:
-            if not html:
-                out = pdoc.render.text(m)
-            else:
-                out = pdoc.render.html_module(
-                    m,
-                    external_links=args.external_links,
-                    link_prefix=args.link_prefix,
-                    source=not args.html_no_source,
-                )
-            print(out, file=w)
-    except Exception:
-        try:
-            os.unlink(f)
-        except:
-            pass
-        raise
-    for submodule in m.submodules():
-        html_out(args, submodule, html)
-
-
 def run():
     """ Command-line entry point """
     args = parser.parse_args()
@@ -215,8 +159,12 @@ def run():
             # Submodules are written to {MODULE_NAME}/{MODULE_NAME}.m.html and
             # subpackages are written to {MODULE_NAME}/{MODULE_NAME}/index.html. And
             # so on...
-            quit_if_exists(args, m)
-            html_out(args, m, args.html)
+            try:
+                pdoc.static.quit_if_exists(args, m)
+                pdoc.static.html_out(args, m, args.html)
+            except pdoc.static.StaticError as e:
+                _eprint(str(e))
+                sys.exit(1)
     else:
         # Plain text
         for m in modules:
